update
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@2751 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/arm/src/lpc17xx/lpc17_internal.h
@@ -537,7 +537,7 @@ EXTERN void lpc17_gpioirqdisable(int irq);
  ************************************************************************************/
 
 #ifdef CONFIG_DEBUG_GPIO
-EXTERN int lpc17_dumpgpio(uint32_t pinset, const char *msg);
+EXTERN int lpc17_dumpgpio(uint16_t pinset, const char *msg);
 #else
 #  define lpc17_dumpgpio(p,m)
 #endif
